requirements.txt
python-dotenv
openai
selenium
python-telegram-bot
psycopg2-binary